break out API handling
